refactor: pull NavigationDrawer into its own component
